+++ tests/pools/selection-strategies/selection-strategies.test.mjs
@@ -277,8 +277,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: {
@@ -333,8 +333,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: {
@@ -559,8 +559,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: {
@@ -618,8 +618,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: {
@@ -758,8 +758,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -827,8 +827,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -977,8 +977,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: {
@@ -1052,8 +1052,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: {
@@ -1208,8 +1208,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -1294,8 +1294,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -1385,8 +1385,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -1586,8 +1586,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -1665,8 +1665,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -1749,8 +1749,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -2012,8 +2012,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({
@@ -2098,8 +2098,8 @@ describe('Selection strategies test suite', () => {
           executing: 0,
           queued: 0,
           maxQueued: 0,
-          stolen: 0,
           sequentiallyStolen: 0,
+          stolen: 0,
           failed: 0
         },
         runTime: expect.objectContaining({